-created class TestTimedTest; - moved tests from TestListTest to TestTutorTest and TestTimedTest; - renamed page TestPage to BaseTestPage and made it abstract - rename SideMenuPage to BaseSideMenu; - use generics for common methods on BaseTestsPage;
