+++ src/test/java/pages/ProfilePage.java
@@ -4,7 +4,7 @@ import com.microsoft.playwright.Locator;
 import com.microsoft.playwright.Page;
 import com.microsoft.playwright.Playwright;
 
-public class ProfilePage extends SideMenuPage {
+public class ProfilePage extends BaseSideMenu {
 
     private final Locator addANewCourseButton = button("Add a new course");
